requirements.txt
streamlit==1.32.0
numpy
pandas
matplotlib
scikit-learn
pillow
opencv-python-headless
imageio
imageio-ffmpeg
tqdm

torch==2.5.1
torchvision==0.20.1
ultralytics==8.4.14